+++ js/main.js
@@ -4,9 +4,7 @@ import './map.js';
 import './mark.js';
 import './similar-marks.js';
 import './offer.js';
-import { similarOffers } from './similar-offers.js';
+import './similar-offers.js';
 import './server.js';
 
-similarOffers.slice();
-
-// console.log(similarOffers);
+// similarOffers.slice();
